% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_blocks_present}
\alias{assert_blocks_present}
\title{Assert that all block features are present in a data.table/data.frame.}
\usage{
assert_blocks_present(colnames_dt, blocks_map, context = "task")
}
\description{
Assert that all block features are present in a data.table/data.frame.
}
\keyword{internal}
