% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_mbspls_state}
\alias{assert_mbspls_state}
\title{Validate the structure of a logged MB-sPLS training snapshot.}
\usage{
assert_mbspls_state(
  st,
  require_train_blocks = FALSE,
  where = "log_env$mbspls_state"
)
}
\description{
Validate the structure of a logged MB-sPLS training snapshot.
}
\keyword{internal}
