% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_flip_weights.R
\name{.mbspls_find_po}
\alias{.mbspls_find_po}
\title{Find first MB-sPLS pipeop in a GraphLearner}
\usage{
.mbspls_find_po(gl)
}
\arguments{
\item{gl}{\link[mlr3pipelines:mlr_learners_graph]{mlr3pipelines::GraphLearner}
A trained graph learner.}
}
\value{
A \link{PipeOpMBsPLS}.
}
\description{
Find first MB-sPLS pipeop in a GraphLearner
}
\keyword{internal}
