% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mbspls.R
\name{impute_knn_graph}
\alias{impute_knn_graph}
\title{Create a k-NN Imputation Graph}
\usage{
impute_knn_graph(k = 5)
}
\arguments{
\item{k}{Integer, number of neighbors (default 5).}
}
\value{
\code{\link[mlr3pipelines:Graph]{mlr3pipelines::Graph}}
}
\description{
k-NN imputation for numeric and factor columns.
}
