% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_align_named_numeric}
\alias{mb_align_named_numeric}
\title{Align a numeric vector to trained feature names.}
\usage{
mb_align_named_numeric(
  v,
  cols,
  context = "vector",
  allow_null = FALSE,
  zero_if_null = FALSE
)
}
\description{
Align a numeric vector to trained feature names.
}
\keyword{internal}
