% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_format_truncated}
\alias{mb_format_truncated}
\title{Format a character vector for concise error messages.}
\usage{
mb_format_truncated(x, max_items = 20L)
}
\description{
Format a character vector for concise error messages.
}
\keyword{internal}
