% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_has_finite_variance}
\alias{mb_has_finite_variance}
\title{Return TRUE only for numeric vectors with finite, positive variance.}
\usage{
mb_has_finite_variance(x, tol = 1e-12)
}
\description{
Return TRUE only for numeric vectors with finite, positive variance.
}
\keyword{internal}
