% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_make_backend_key_name}
\alias{mb_make_backend_key_name}
\title{Create a backend primary-key column name that does not collide.}
\usage{
mb_make_backend_key_name(existing, key_name = "..row_id")
}
\description{
Create a backend primary-key column name that does not collide.
}
\keyword{internal}
