% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_normalize_blocks}
\alias{mb_normalize_blocks}
\title{Normalize a multi-block mapping.}
\usage{
mb_normalize_blocks(blocks, .var.name = "blocks")
}
\description{
Normalize a multi-block mapping.
}
\keyword{internal}
