% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_resolve_blocks}
\alias{mb_resolve_blocks}
\title{Resolve blocks against a concrete data table.}
\usage{
mb_resolve_blocks(dt, blocks, numeric_only = TRUE, non_constant = TRUE)
}
\description{
Resolve blocks against a concrete data table.
}
\keyword{internal}
