% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_task_blocks}
\alias{mb_task_blocks}
\title{Extract block metadata from a multiblock task if available.}
\usage{
mb_task_blocks(task, context = "task", allow_null = FALSE)
}
\description{
Extract block metadata from a multiblock task if available.
}
\keyword{internal}
