% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_validate_supervised_task}
\alias{mb_validate_supervised_task}
\title{Validate supervised TaskMultiBlock usage.}
\usage{
mb_validate_supervised_task(task, context = "mbsplsxy_graph")
}
\description{
Validate supervised TaskMultiBlock usage.
}
\keyword{internal}
