% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mbspls.R
\name{MeasureMBsPLS_EV}
\alias{MeasureMBsPLS_EV}
\title{MB-sPLS mean explained variance across components (prediction-side)}
\description{
Mean of per-component explained variance on the prediction split.
Incremental out-of-sample EV can be negative, so this measure is unbounded
below and above in principle. Scoring requires stored models; see
\link{mbspls_measures}.
}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureMBsPLS_EV}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureMBsPLS_EV-new}{\code{MeasureMBsPLS_EV$new()}}
\item \href{#method-MeasureMBsPLS_EV-clone}{\code{MeasureMBsPLS_EV$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="obs_loss"><a href='../../mlr3/html/Measure.html#method-Measure-obs_loss'><code>mlr3::Measure$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureMBsPLS_EV-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureMBsPLS_EV-new}{}}}
\subsection{Method \code{new()}}{
Construct the measure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMBsPLS_EV$new(id = "mbspls.ev")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character(1). Measure ID (default shown).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureMBsPLS_EV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureMBsPLS_EV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMBsPLS_EV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
