% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_expand_block_cols}
\alias{mb_expand_block_cols}
\title{Expand stable base names to concrete task/backend column names.}
\usage{
mb_expand_block_cols(dt_names, cols)
}
\description{
Expand stable base names to concrete task/backend column names.
}
\keyword{internal}
