% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mbspls.R
\name{mbspls_graph_learner}
\alias{mbspls_graph_learner}
\title{MB‑sPLS GraphLearner: preproc → MB‑sPLS → bootstrap‑select (two selection methods) → learner}
\usage{
mbspls_graph_learner(
  learner = lrn("clust.kmeans", centers = 1L),
  blocks,
  site_correction,
  site_correction_methods,
  ncomp,
  ...
)
}
\arguments{
\item{learner}{Downstream learner (default k‑means with 1 center).}

\item{blocks, site_correction, site_correction_methods, ncomp}{Main settings.}

\item{...}{Additional arguments passed to \code{\link[=mbspls_graph]{mbspls_graph()}}.}
}
\value{
\link[mlr3pipelines:mlr_learners_graph]{mlr3pipelines::GraphLearner}
}
\description{
MB‑sPLS GraphLearner: preproc → MB‑sPLS → bootstrap‑select (two selection methods) → learner
}
