% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mbspls.R
\name{mbspls_graph_learner}
\alias{mbspls_graph_learner}
\title{MB-sPLS GraphLearner: preproc -> MB-sPLS -> bootstrap-select (two selection methods) -> learner}
\usage{
mbspls_graph_learner(learner = lrn("clust.kmeans", centers = 1L), ...)
}
\arguments{
\item{learner}{Downstream learner (default k-means with 1 center).}

\item{...}{Arguments passed to \code{\link[=mbspls_graph]{mbspls_graph()}}.}
}
\value{
\link[mlr3pipelines:mlr_learners_graph]{mlr3pipelines::GraphLearner}
}
\description{
MB-sPLS GraphLearner: preproc -> MB-sPLS -> bootstrap-select (two selection methods) -> learner
}
