% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mbspls.R
\name{mbspls_preproc_graph}
\alias{mbspls_preproc_graph}
\title{Default MB‑sPLS Preprocessing Graph}
\usage{
mbspls_preproc_graph(
  blocks,
  site_correction,
  site_correction_methods,
  keep_site_col = FALSE,
  k = 5,
  id_suffix = NULL
)
}
\arguments{
\item{blocks}{Named list of character vectors per block.}

\item{site_correction}{Named list of features used for site correction.}

\item{site_correction_methods}{Named list of methods for site correction.}

\item{keep_site_col}{Keep site column after correction?}

\item{k}{k for kNN imputation (default 5).}

\item{id_suffix}{Optional suffix for PipeOp ids.}
}
\value{
\code{\link[mlr3pipelines:Graph]{mlr3pipelines::Graph}}
}
\description{
Site correction → encoding/imputation → scaling. No MB‑sPLS here.
}
