% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mbspca.R
\name{MeasureMBSPCAMEV}
\alias{MeasureMBSPCAMEV}
\title{MB-SPCA Mean Explained Variance}
\description{
A custom \link[mlr3:Measure]{mlr3::Measure} that computes the mean explained variance of the latent variables in MB-SPCA.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{MeasureMBSPCAMEV$new()
}\if{html}{\out{</div>}}
}

\section{Methods}{

\itemize{
\item \verb{$score(prediction, task, learner, ...)}: Computes the measure based on the prediction and task.
}
}

\examples{
\dontrun{
measure = MeasureMBSPCAMEV$new()
}

}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureMBSPCAMEV}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureMBSPCAMEV-new}{\code{MeasureMBSPCAMEV$new()}}
\item \href{#method-MeasureMBSPCAMEV-clone}{\code{MeasureMBSPCAMEV$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureMBSPCAMEV-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureMBSPCAMEV-new}{}}}
\subsection{Method \code{new()}}{
Construct the measure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMBSPCAMEV$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character(1). Measure ID (default shown).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureMBSPCAMEV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureMBSPCAMEV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMBSPCAMEV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
