% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mbspls.R
\name{MeasureMBsPLS_EVWeightedMAC}
\alias{MeasureMBsPLS_EVWeightedMAC}
\title{MB-sPLS EV-weighted latent correlation (test)}
\description{
Aggregates per-component test latent correlations (MAC or Frobenius) using
explained-variance weights from the same test split:
\deqn{\sum_k \left(\frac{EV_k}{\sum_j EV_j}\right) \cdot \text{score}_k.}
If Frobenius was used during training, scores are normalised by
\eqn{\sqrt{\binom{B}{2}}} to keep the range within \eqn{[0,1]}.
}
\examples{
\dontrun{
msr_evwt = MeasureMBsPLS_EVWeightedMAC$new()
}
}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureMBsPLS_EVWeightedMAC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureMBsPLS_EVWeightedMAC-new}{\code{MeasureMBsPLS_EVWeightedMAC$new()}}
\item \href{#method-MeasureMBsPLS_EVWeightedMAC-clone}{\code{MeasureMBsPLS_EVWeightedMAC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="obs_loss"><a href='../../mlr3/html/Measure.html#method-Measure-obs_loss'><code>mlr3::Measure$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureMBsPLS_EVWeightedMAC-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureMBsPLS_EVWeightedMAC-new}{}}}
\subsection{Method \code{new()}}{
Construct the measure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMBsPLS_EVWeightedMAC$new(id = "mbspls.mac_evwt")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character(1). Measure ID (default shown).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureMBsPLS_EVWeightedMAC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureMBsPLS_EVWeightedMAC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMBsPLS_EVWeightedMAC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
